[level 1] Title: 콜라츠 추측, Time: 0.07 ms, Memory: 87.5 MB -BaekjoonHub

+++ 프로그래머스/1/12943. 콜라츠 추측/콜라츠 추측.java
@@ -1,7 +1,7 @@
 class Solution {
     public int solution(int num) {
         int answer = 0;//몇번하는지 횟수
-        double tmp = num;
+        long tmp = num;
         while(tmp!=1){
             if(tmp%2==0){
                 tmp=tmp/2;
